attr: use SHELL from configure
If /bin/sh is not a functional enough shell, configure will select a
SHELL of /bin/bash or better.  But the current build helpers always
hardcode /bin/sh, so if libtool itself configures itself for /bin/bash,
things will fail when it attempts to do:
	SHELL = /bin/sh
	LIBTOOL = $(SHELL) .../libtool
	...
	eval: 1: base_compile+= -pipe: not found
	...

So rather than hardcoding SHELL to /bin/sh, set it to @SHELL@ and let
configure find a good value for us.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>