test: make running parallel/out-of-tree safe
Change the run program to update PATH to find the tools locally
automatically.  This lets us avoid having to always build in the
tree and allows for direct running.

Also update the program to create a tempdir for each test to run
inside of.  This way we can run all the tests in parallel without
randomly clobbering each other.