modernize build system
This deletes the hand rolled build system and replaces it entirely
with autotools.  The overall diffstat shows that this is a clear
win, and it makes the package build/install like every other autotool
package out there which makes the lives of distro maintainers a lot
easier.

This should also be faster by virtue of using a non-recursive build.

Things to note:
 - to generate autotools:
	./autogen.sh
 - to see full compile output:
	make V=1
 - to build specific targets:
	make attr getfattr ...
 - to run tests:
	make check
 - to create a release:
	make distcheck -j

Other non-developer things:
 - the xattr.conf file is now installed into sysconfdir by default
 - man pages are no longer compressed as this is uncommon in the
   autotools world and distros don't need it anyways (they already
   handle it automatically for most pakages)
 - the minor # of the shared library is now based on the package
   version so it'll be much bigger ... this isn't a problem, and
   is actually a bugfix (older releases didn't change when they
   should have)

(Again, this is all using the standard autotool targets.)