test: fix cleanup & running as root
This fixes a bug when doing the cleanup of tests in a subdir by making
sure we chdir back to our starting point.

We also move the root test out of XFAIL and instead do the root checking
at runtime.  If we are root, the test passes nicely, otherwise it marks
itself as SKIP quickly.

These are fixing issues noticed when doing the update to libacl.