libattr: Set symbol versions for legacy syscalls via attribute or asm
Currently, a linker script is used to define versioned symbols
for the legacy linux syscalls with simple assignments.

That isn't mentioned in ld's documentation as a valid method to
set symbol versions, and while one might assume that it should work
and the linker script is accepted, the result often isn't correct:
gold and lld always create broken binaries, and even the bfd linker
can create unusable symbols if used with --gc-sections or LTO.

For example, the result can be a library where the function has been
discarded and the versioned symbol is unusable, i.e.,
    23: 00000000     0 NOTYPE  GLOBAL DEFAULT  ABS getxattr@ATTR_1.0
instead of
    23: 000033c0     0 FUNC    GLOBAL DEFAULT   11 getxattr@ATTR_1.0
Meanwhile, lld doubles the version suffix: getxattr@ATTR_1.0@ATTR_1.0

Not that these issues may be viewed as linker bugs, and there may
or may not be some related improvements to linker script handling
since I tested this a few months ago (binutils-2.37, lld-14).
But in either case a more robust solution would be preferable.

So remove the linker script entirely and set symbol versions with
__attribute__((__symver__(...))) if available (i.e., in gcc >= 10,
but not in clang).  Otherwise use the traditional global asm solution
with a .symver directive.  These are the well documented methods (along
with version scripts, which don't apply here) to set symbol versions
and work correctly with all linkers, including older versions.

A workaround that is needed for gcc < 10 not to break LTO partitioning
with global asm is also included (__attribute__((__no_reorder__))),
but some very old versions may still need -flto-partition=none to build
correctly with LTO (not going to bother with completely obsolete versions).

Signed-off-by: Alexander Miller <alex.miller@gmx.de>