configure.ac: add detection of symver gcc attribute
On non-ELF platforms, such as microblaze, builds will fail when trying
to add symver information because  __attribute__((symver ..)) is not
supported even though __has_attribute(__symver__) returns true.

Support for symver needs to be detected via a compile test since
__has_attribute can report false positives [0].

Add a configure compile check for __attribute__((symver ..)) to ensure
it is supported and define a variable to advertise support.

[0] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=101766#c1

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
[Giulio: rework #if nesting]
Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>