Fix listxattr-related races and stack overflows
This should fix several problems that I noticed in libxattr
when testing bleeding-edge coreutils (Savannah commit
4eb2ff1c8984403e7f7def1cf22eab0b0e3c3dad dated 2025-05-29)
on Fedora 42 x86-64 (Linux kernel 6.14.8-300.fc42.x86_64).
The problems were:
- race condition if some other process adds attributes
- potentially unbounded stack allocation
- this kernel has a bug where flistxattr needs to be
  passed a larger size than what is actually needed
- unnecessary syscall in the usual case
- unnecessarily duplicated error-handling code
* configure.ac (AC_FUNC_ALLOCA): Remove; no longer needed.
* libattr/attr_copy_fd.c, libattr/attr_copy_file.c:
Do not include alloca.h.
(my_alloc, my_free): Remove.  All uses replaced by malloc, free.
(attr_copy_fd, attr_copy_file): Do not rely on passing NULL to get a
reliable size, as there may be a race if some other process is adding
attributes, and anyway the NULL-derived size is not enough for Fedora 42
as of 2025-05-30 due to a Linux kernel bug which requires you to pass a
larger size than needed in a tmpfs file.  Also, use a small buffer on
the stack for the first call; this avoids a syscall and a malloc or
alloca in the usual case where only a small buffer is needed.
Always do later allocations on the heap to avoid potential stack
overflow bugs with alloca.  Simplify by coalescing duplicate calls to
quote, error, quote_free.