Fix lgetxattr-related races
This fixes some races related to fgetxattr and
lgetxattr, that were similar to the recently
fixed listxattr-related races.
* libattr/attr_copy_fd.c (attr_copy_fd):
* libattr/attr_copy_file.c (attr_copy_file):
Fix race if an attribute changes between getting
its value's size and getting the actual value.
Do not bother to query the attribute value size,
to avoid an unnecessary syscall in the usual case.
Just call malloc+free, not realloc, as there is
no need to preserve the old data.